requirements.txt
azure-storage-blob==12.20.0
python-dotenv==1.0.1
tqdm==4.66.4
